rename to feature package
