+++ src/main/java/org/betonquest/betonquest/api/quest/npc/feature/NpcInteractCatcher.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.api.quest.npc.conversation;
+package org.betonquest.betonquest.api.quest.npc.feature;
 
 import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.api.NpcInteractEvent;
